Editor Revisions List: Split list items into their own components
